+++ include/simplex_subset.h
@@ -5,16 +5,16 @@
 class SimplexSubset {
 
   public:
-    std::set<size_t> vertices;
-    std::set<size_t> edges;
-    std::set<size_t> faces;
-    std::set<size_t> tets;
+    std::set<int> vertices;
+    std::set<int> edges;
+    std::set<int> faces;
+    std::set<int> tets;
 
     /* Initialize an empty SimplexSubset. */
     SimplexSubset() {}
     
     /* Initialize a SimplexSubset with the given vertices, edges, and faces. */
-    SimplexSubset(const std::set<size_t>& V, const std::set<size_t>& E, const std::set<size_t>& F) {
+    SimplexSubset(const std::set<int>& V, const std::set<int>& E, const std::set<int>& F) {
 
         vertices = V;
         edges = E;
@@ -22,7 +22,7 @@ class SimplexSubset {
     }
     
     /* Initialize a SimplexSubset with the given vertices, edges, and faces. */
-    SimplexSubset(const std::set<size_t>& V, const std::set<size_t>& E, const std::set<size_t>& F, const std::set<size_t>& T) {
+    SimplexSubset(const std::set<int>& V, const std::set<int>& E, const std::set<int>& F, const std::set<int>& T) {
 
         vertices = V;
         edges = E;
@@ -34,27 +34,27 @@ class SimplexSubset {
      * SimplexSubset.
      */
     SimplexSubset deepCopy() const {
-        std::set<size_t> newVertices = vertices;
-        std::set<size_t> newEdges = edges;
-        std::set<size_t> newFaces = faces;
-        std::set<size_t> newTets = tets;
+        std::set<int> newVertices = vertices;
+        std::set<int> newEdges = edges;
+        std::set<int> newFaces = faces;
+        std::set<int> newTets = tets;
         return SimplexSubset(newVertices, newEdges, newFaces, newTets);
     }
 
     /* Add a vertex to this subset. */
-    void addVertex(size_t index) {
+    void addVertex(int index) {
         vertices.insert(index);
     }
 
     /* Add a set of vertices to this subset. */
-    void addVertices(const std::set<size_t>& V) {
-        for (std::set<size_t>::iterator it = V.begin(); it != V.end(); ++it) {
+    void addVertices(const std::set<int>& V) {
+        for (std::set<int>::iterator it = V.begin(); it != V.end(); ++it) {
             vertices.insert(*it);
         }
     }
 
     /* Delete a vertex from this subset. */
-    void deleteVertex(size_t index) {
+    void deleteVertex(int index) {
         vertices.erase(index);
     }
 
@@ -64,26 +64,26 @@ class SimplexSubset {
     }
 
     /* Delete a set of vertices from this subset. */
-    void deleteVertices(const std::set<size_t>& V) {
-        for (std::set<size_t>::iterator it = V.begin(); it != V.end(); ++it) {
+    void deleteVertices(const std::set<int>& V) {
+        for (std::set<int>::iterator it = V.begin(); it != V.end(); ++it) {
             vertices.erase(*it);
         }
     }
 
     /* Add an edge to this subset. */
-    void addEdge(size_t index) {
+    void addEdge(int index) {
         edges.insert(index);
     }
 
     /* Add a set of edges to this subset. */
-    void addEdges(const std::set<size_t>& E) {
-        for (std::set<size_t>::iterator it = E.begin(); it != E.end(); ++it) {
+    void addEdges(const std::set<int>& E) {
+        for (std::set<int>::iterator it = E.begin(); it != E.end(); ++it) {
             edges.insert(*it);
         }
     }
 
     /* Delete an edge from this subset. */
-    void deleteEdge(size_t index) {
+    void deleteEdge(int index) {
         edges.erase(index);
     }
 
@@ -93,26 +93,26 @@ class SimplexSubset {
     }
 
     /* Delete a set of edges from this subset. */
-    void deleteEdges(const std::set<size_t>& E) {
-        for (std::set<size_t>::iterator it = E.begin(); it != E.end(); ++it) {
+    void deleteEdges(const std::set<int>& E) {
+        for (std::set<int>::iterator it = E.begin(); it != E.end(); ++it) {
             edges.erase(*it);
         }
     }
     
     /* Add a face to this subset. */
-    void addFace(size_t index) {
+    void addFace(int index) {
         faces.insert(index);
     }
 
     /* Add a set of faces to this subset. */
-    void addFaces(const std::set<size_t>& F) {
-        for (std::set<size_t>::iterator it = F.begin(); it != F.end(); ++it) {
+    void addFaces(const std::set<int>& F) {
+        for (std::set<int>::iterator it = F.begin(); it != F.end(); ++it) {
             faces.insert(*it);
         }
     }
 
     /* Delete a face from this subset. */
-    void deleteFace(size_t index) {
+    void deleteFace(int index) {
         faces.erase(index);
     }
 
@@ -122,26 +122,26 @@ class SimplexSubset {
     }
 
     /* Delete a set of faces from this subset. */
-    void deleteFaces(const std::set<size_t>& F) {
-        for (std::set<size_t>::iterator it = F.begin(); it != F.end(); ++it) {
+    void deleteFaces(const std::set<int>& F) {
+        for (std::set<int>::iterator it = F.begin(); it != F.end(); ++it) {
             faces.erase(*it);
         }
     }
     
     /* Add a tet to this subset. */
-    void addTet(size_t index) {
+    void addTet(int index) {
         tets.insert(index);
     }
 
     /* Add a set of tets to this subset. */
-    void addTets(const std::set<size_t>& T) {
-        for (std::set<size_t>::iterator it = T.begin(); it != T.end(); ++it) {
+    void addTets(const std::set<int>& T) {
+        for (std::set<int>::iterator it = T.begin(); it != T.end(); ++it) {
             tets.insert(*it);
         }
     }
 
     /* Delete a tet from this subset. */
-    void deleteTet(size_t index) {
+    void deleteTet(int index) {
         tets.erase(index);
     }
 
@@ -151,8 +151,8 @@ class SimplexSubset {
     }
 
     /* Delete a set of tets from this subset. */
-    void deleteTets(const std::set<size_t>& T) {
-        for (std::set<size_t>::iterator it = T.begin(); it != T.end(); ++it) {
+    void deleteTets(const std::set<int>& T) {
+        for (std::set<int>::iterator it = T.begin(); it != T.end(); ++it) {
             tets.erase(*it);
         }
     }
@@ -179,37 +179,37 @@ class SimplexSubset {
         this->deleteTets(other.tets);
     }
 
-    /* Prsize_t vertices. */
+    /* Print vertices. */
     void printVertices() {
         std::cout << "Vertices: ";
-        for (std::set<size_t>::iterator it = vertices.begin(); it != vertices.end(); ++it) {
+        for (std::set<int>::iterator it = vertices.begin(); it != vertices.end(); ++it) {
             std::cout << *it << ", ";
         }
         std::cerr << std::endl;
     }
 
-    /* Prsize_t edges. */
+    /* Print edges. */
     void printEdges() {
         std::cout << "Edges: ";
-        for (std::set<size_t>::iterator it = edges.begin(); it != edges.end(); ++it) {
+        for (std::set<int>::iterator it = edges.begin(); it != edges.end(); ++it) {
             std::cout << *it << ", ";
         }
         std::cerr << std::endl;
     }
     
-    /* Prsize_t Faces. */
+    /* Print Faces. */
     void printFaces() {
         std::cout << "Faces: ";
-        for (std::set<size_t>::iterator it = faces.begin(); it != faces.end(); ++it) {
+        for (std::set<int>::iterator it = faces.begin(); it != faces.end(); ++it) {
             std::cout << *it << ", ";
         }
         std::cerr << std::endl;
     }
     
-    /* Prsize_t Tets. */
+    /* Print Tets. */
     void printTets() {
         std::cout << "Tets: ";
-        for (std::set<size_t>::iterator it = tets.begin(); it != tets.end(); ++it) {
+        for (std::set<int>::iterator it = tets.begin(); it != tets.end(); ++it) {
             std::cout << *it << ", ";
         }
         std::cerr << std::endl;
@@ -222,7 +222,7 @@ class SimplexSubset {
         deleteTets();
     }
     
-    /* Prsize_t All. */
+    /* Print All. */
     void printAll() {
         printTets();
         printFaces();
